requirements.txt
pandas==2.3.1
psycopg2-binary==2.9.9
dotenv==0.9.9
sentence-transformers==5.0.0
torch==2.7.1